refactor: code refactor of cluster, search api
