+++ src/App.js
@@ -1,5 +1,5 @@
 import "./App.css";
-import mainRouter from "./main/pages/router";
+import mainRouter from "./main/router/router";
 import { RouterProvider } from "react-router-dom";
 
 function App() {
