+++ src/navbar/components/styles/style.css
@@ -8,26 +8,41 @@
   max-height: 100px;
 }
 
+.searchForm .form-control::placeholder {
+  font-size: 0.9em; /* 원하는 크기로 조정 */
+  color: #6c757d; /* 선택 사항: placeholder 글자 색상 조정 */
+}
+
 .searchForm {
   display: flex;
   flex-direction: column;
   position: relative;
+  width: 100%;
 }
 
 .includedSearchForm {
   position: absolute;
   top: 100%;
   left: 0;
-  z-index: 1;
+  right: 0;
+  z-index: 11;
   display: flex;
   flex-direction: column;
-  border: 1px solid #000000;
   background-color: white;
   border-radius: 0 0 10px 10px;
   padding: 10px;
+  box-sizing: border-box; 
+  border: 1px solid #ccc;
+  overflow-y: auto;
 }
 
-.includedSearchForm a {
+.included-result {
   text-decoration: none;
   color: black;
+  padding: 5px 0;
+  cursor: pointer;
+}
+
+.included-result:hover {
+  background-color: #f1f1f1;
 }
